simplify context hook impl
